Move tests to /tests
