+++ package.json
@@ -8,9 +8,9 @@
     "start": "node ./lib/app.js",
     "build": "tsc",
     "test": "npm run test:unit && npm run test:integration",
-    "test:unit": "mocha --reporter list --ui bdd --require ts-node/register --require source-map-support/register --recursive lib/tests/unit",
-    "test:integration": "mocha --exit --reporter list --ui bdd --require ts-node/register --require source-map-support/register --recursive lib/tests/integration",
-    "test:postgres": "SLACKBRIDGE_TEST_ENABLEPG=yes mocha --reporter list --ui bdd --require ts-node/register --require source-map-support/register --recursive lib/tests/integration/PgDatastoreTest.js",
+    "test:unit": "mocha --reporter list --ui bdd --require ts-node/register --require source-map-support/register --recursive tests/unit/*.ts",
+    "test:integration": "mocha --exit --reporter list --ui bdd --require ts-node/register --require source-map-support/register --recursive tests/integration/*.ts",
+    "test:postgres": "SLACKBRIDGE_TEST_ENABLEPG=yes mocha --reporter list --ui bdd --require ts-node/register --require source-map-support/register --recursive tests/integration/PgDatastoreTest.ts",
     "lint": "eslint -c .eslintrc.js ./src"
   },
   "repository": {
