+++ tests/unit/BridgedRoomTest.ts
@@ -1,0 +1,27 @@
+/*
+Copyright 2019 The Matrix.org Foundation C.I.C.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+import { BridgedRoom } from "../../src/BridgedRoom";
+
+describe("BridgedRoom", () => {
+    it("constructs", () => {
+        new BridgedRoom({} as any, {
+            inbound_id: "123456a",
+            matrix_room_id: "!abcde:localhost",
+            slack_type: "unknown",
+        });
+    });
+});
