+++ tests/utils/slackTestApi.ts
@@ -1,7 +1,7 @@
 import { WebClientOptions } from "@slack/web-api";
 import { createServer, Server, IncomingMessage, ServerResponse } from "http";
 import { promisify } from "util";
-import { TeamInfoResponse, AuthTestResponse, UsersInfoResponse } from "../../SlackResponses";
+import { TeamInfoResponse, AuthTestResponse, UsersInfoResponse } from "../../src/SlackResponses";
 import * as qs from "querystring";
 
 export class SlackTestApi {
