+++ tsconfig.json
@@ -14,5 +14,9 @@
   },
   "include": [
     "src/**/*"
+  ],
+  "exclude": [
+    "tests/**/*",
   ]
+
 }
